+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Introduce CUBE CUT. Press C in edit mode to enter cube cut mode. See Readme for more details.
 - Fix grids not displaying in orthographic view in some cases
 - On Level Design workspace creation, set appropriate blender file settings
+- Set paths to relative on save
 
 ## 1.0.3